+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.06] - 2025-2-19
+
+### Added
+
+- Add twilio to the creation process
+- Add phone registration to the creation process (step 4)
+
+### Changed
+
+- Change SelectionMenu to be scrollable
+
 ## [0.4.05] - 2025-2-17
 
 ### Added